frame generator fix
The code in frame_generator.py generates a command for ffmpeg
to cut out a section of the display, and then it forwards the
video in this section of the screen to the Trinus app on the
phone.

The code is hardcoded to look on the first VT, which is
sometimes OK, but probably not on some linuxen, including
Ubuntu. Let's switch it to use the value in DISPLAY, which is
set by the user's session to the correct VT.

Signed-off-by: Benjamin M Romer <benjamin.romer@canonical.com>