initial makefile
I'd like to package this up for Ubuntu as a deb, so it will
need to be able to build in a builder.

Also, new people have a lot of trouble getting the driver to
work, so the makefile can serve as a hint for what is needed
to make the application work.

Signed-off-by: Benjamin M Romer <benjamin.romer@canonical.com>